Execute a list of strings rather than one big string
While vim v8.0.654 made the "one big string" approach possible, this
version is still new enough that you basically need to either be
compiling your own vim, using a PPA (Ubuntu), or running archlinux.

Personally, I prefer the archlinux route, but that's not always
possible.  (Or desirable, depending on the person.)

With this change, our minimum working version of vim is v7.4.1304.