Set up a `local::lib` in `perl5/`
...because it's not unlikely we'd want to install additional CPAN
modules somewhere, and we may as well not clutter the system.

We probably need to think about this a bit more and revisit down the
line, as setting a `local::lib` here will change it for spawned
processes as well.  (good when we invoke `cpanm`, potentially bad when
`$USER` invokes something)