"Fake" a `local::lib` by default
`local::lib` is nifty, but while we're using it to provide modules for
us to use within vim it has a global effect for anything invoked by vim.

Which, as you can imagine, is not generally what one wants.

So, by default we now fake it, by adding the expected entries to `@INC`
and using `local::lib` itself to ensure the directory structure is
correctly created.  This won't impact functions like
`ducttape#install()`, as they invoke `cpanm` with the target explicitly
defined via the `--local-lib` option.